Separate styles from screens
