+++ screens/Camera/index.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useRef } from 'react';
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
+import { Text, View, TouchableOpacity } from 'react-native';
 import { Camera } from 'expo-camera';
-import colors from '../theme/color';
+import styles from './styles';
 
 export default function Picture({ route, navigation }) {
   const cameraRef = useRef(null);
@@ -58,48 +58,3 @@ export default function Picture({ route, navigation }) {
     </>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flexGrow: 0.8,
-    justifyContent: 'center',
-    alignItems: 'center'
-  },
-  camera: {
-    flex: 1,
-    width: '100%'
-  },
-  buttonContainer: {
-    flexGrow: 0.2,
-    backgroundColor: 'transparent',
-    flexDirection: 'row',
-    justifyContent: 'center',
-    margin: 20,
-  },
-  button: {
-    flex: 0.1,
-    alignSelf: 'flex-end',
-    alignItems: 'center',
-  },
-  snap: {
-    width: 100,
-    height: 100,
-    borderRadius: 100,
-    alignSelf: 'center',
-    backgroundColor: colors.main
-  },
-  text: {
-    fontSize: 18,
-    color: 'white',
-  },
-  taken: {
-    width: 100,
-    height: 180,
-    backgroundColor: 'white'
-  },
-  stretch: {
-    width: 120,
-    height: 180,
-    resizeMode: 'stretch',
-  },
-});
