+++ screens/MyPage/styles.js
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center"
+  },
+  info: {
+    width: '60%',
+    marginTop: 42
+  },
+  button: {
+    width: '60%',
+    alignSelf: 'center'
+  }
+});
+
+export default styles;
