+++ screens/MySubmissions/index.js
@@ -1,15 +1,21 @@
 import React, { useState, useCallback } from 'react';
-import { useFocusEffect } from '@react-navigation/native';
 import { useSelector, useDispatch } from 'react-redux';
-import { createError } from '../features/userSlice';
-import errorMessage from '../constants/errorMessage';
+import { useFocusEffect } from '@react-navigation/native';
+import { createError } from '../../features/userSlice';
+import errorMessage from '../../constants/errorMessage';
 
-import { View, Text, Image, FlatList, StyleSheet } from 'react-native';
-import Empty from '../components/shared/Empty';
+import {
+  View,
+  Text,
+  Image,
+  FlatList
+} from 'react-native';
+import Empty from '../../components/shared/Empty';
+import styles from './styles';
 
-import submissionAPI from '../api/submissions';
-import { formatDate } from '../utils/index';
-import colors from '../theme/color';
+import submissionAPI from '../../api/submissions';
+import { formatDate } from '../../utils/index';
+
 
 const MySubmissions = () => {
   const dispatch = useDispatch();
@@ -110,56 +116,5 @@ const MySubmissions = () => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    margin: 10,
-    paddingTop: 12,
-    paddingBottom: 10,
-  },
-  title: {
-    alignSelf: 'center',
-    fontSize: 16,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  submissionContainer: {
-    flexDirection: 'row',
-    backgroundColor: colors.white,
-    marginTop: 10,
-    padding: 10,
-    borderRadius: 14,
-  },
-  leftContainer: {
-    flexDirection: 'row',
-    flex: 5,
-    justifyContent: 'space-around'
-  },
-  rightContainer: {
-    flex: 1,
-  },
-  image: {
-    flex: 1,
-    width: '10%',
-    height: undefined,
-    aspectRatio: 1,
-    borderRadius: 56,
-    marginLeft: 6,
-  },
-  messageContainer: {
-    flex: 4,
-    marginLeft: 10,
-  },
-  name: {
-    fontWeight: 'bold',
-    fontSize: 16
-  },
-  message: {
-    fontSize: 14
-  },
-  icon: {
-    alignSelf: 'center'
-  }
-});
 
 export default MySubmissions;
