+++ screens/Signup/styles.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  screen: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+    paddingTop: 50,
+  },
+  title: {
+    fontSize: 60,
+  },
+  logo: {
+    width: 150,
+    height: 150,
+  },
+  inputContainer: {
+    width: 300,
+  },
+});
+
+export default styles;
